pyproject.toml
[project]
name = "rpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hemac",
    "pyyaml>=6.0.3",
]

[tool.uv.workspace]
members = [
    "hemac",
]

[tool.uv.sources]
hemac = { workspace = true, editable = true }
